Create utils/configuration_backup module
